Some basic tests for LBAccount.
Uncommented -sentDate in LBMessage.m however there is a preexisting comment that it might not be the best way to calculate the sent date so it should be looked at sometime.

LBMIMETests.m used an absolute path for the testdata folder. I switched it to a relative path, but it might not work for everyone... However since the tests aren't run unless you explicitly run them it's not a big concern right now.